+++ Service/src/main/java/org/gusdb/wdk/service/formatter/UserDatasetFormatter.java
@@ -124,7 +124,7 @@ public class UserDatasetFormatter {
       }
 
       json.put("sharedWith", sharesJson);
-      int quota = datasetInfo.getOwnerQuota();
+      long quota = datasetInfo.getOwnerQuota();
       DecimalFormat df = new DecimalFormat("#.####");
       json.put("percentQuotaUsed", df.format(dataset.getSize() * 100.0 / quota));
     }
